Batching in to_glow compilation
Summary:
Add a loop for seq_len and batch_size, also factor out this code into another file

Note: There's an additional diff which is not in this stack which is needed (D24760292). I have it in my local stack just above D24553719, and it required some manual merging to get it there.

Differential Revision: D24923845

fbshipit-source-id: 02d2629f889e780f6402ffcfcd03d0adde4e2186